fastapi_example_cdk/requirements.txt
aws-cdk-lib==2.57.0
constructs>=10.0.0,<11.0.0
#aws-cdk.aws-apigatewayv2-alpha
#aws-cdk.aws-apigatewayv2-integrations-alpha
aws-cdk.aws-lambda-python-alpha